파일 매니저 코어 분리
