+++ web/src/core/file-manager/FileManager.ts
@@ -1,0 +1,9 @@
+export class FileManager {
+  constructor(private fileSystem: FileSystem) {}
+
+  async getDirectory(path: string, options?: Flags) {
+    return new Promise<DirectoryEntry>((res, rej) =>
+      this.fileSystem.root.getDirectory(path, options, res, rej)
+    )
+  }
+}
