+++ web/src/core/file-manager/index.ts
@@ -1,0 +1,2 @@
+export * from './configureFileManager'
+export * from './FileManager'
