+++ web/src/global.d.ts
@@ -1,0 +1,9 @@
+interface Navigator {
+  webkitPersistentStorage: {
+    requestQuota(
+      size: number,
+      onSuccess: (size: number) => void,
+      onFailure?: (error: DOMError) => void
+    ): void
+  }
+}
